[MOD-9694] Trie prefixes iterator (#6119)
* Prefixes iterator

* Align with C: do not track keys in the prefixes iterator

* Compile arr.h as a standalone static library

* Fix memory issue in bench

* Add license header

* Re-add tests for values iterator

* Add another prefix benchmark using the full Gutenberg dataset.
