+++ src/redisearch_rs/trie_bencher/benches/iter.rs
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+//! Benchmark the iterators exposed by a trie map.
+
+use criterion::{Criterion, criterion_group, criterion_main};
+use trie_bencher::OperationBencher;
+
+use trie_bencher::corpus::CorpusType;
+
+fn iter_benches_wiki1k(c: &mut Criterion) {
+    let corpus = CorpusType::RedisBench1kWiki;
+    let terms = corpus.create_terms(true);
+
+    let bencher = OperationBencher::new("Wiki-1K".to_owned(), terms, None);
+    // Matches "A" and "Abacus"
+    bencher.find_prefixes_group(c, "Abacuses", "Find prefixes");
+}
+
+fn iter_benches_gutenberg(c: &mut Criterion) {
+    let corpus = CorpusType::GutenbergEbook(true);
+    let terms = corpus.create_terms(true);
+
+    let bencher = OperationBencher::new("Gutenberg".to_owned(), terms, None);
+    // Matches "ever", "everlasting" and "everlastingly".
+    bencher.find_prefixes_group(c, "everlastingly", "Find prefixes");
+}
+
+criterion_group!(wiki_1k_iter, iter_benches_wiki1k);
+criterion_group!(gutenberg_iter, iter_benches_gutenberg);
+criterion_main!(wiki_1k_iter, gutenberg_iter);
