+++ src/redisearch_rs/trie_bencher/src/corpus.rs
@@ -207,9 +207,12 @@ impl CorpusType {
     fn get_pretty_print_path(&self) -> PathBuf {
         let path = PathBuf::from("data");
         let filename = match self {
-            CorpusType::RedisBench1kWiki => "redis_wiki1k_titles_bench.txt",
-            CorpusType::RedisBench10kNumerics => "redis_wiki10k_guids_bench.txt",
-            CorpusType::GutenbergEbook(_) => "gutenberg_bench.txt",
+            CorpusType::RedisBench1kWiki => "redis_wiki1k_titles_bench.txt".to_owned(),
+            CorpusType::RedisBench10kNumerics => "redis_wiki10k_guids_bench.txt".to_owned(),
+            CorpusType::GutenbergEbook(full) => {
+                let suffix = if *full { "full" } else { "sample" };
+                format!("gutenberg_bench_{suffix}.txt")
+            }
         };
         path.join(filename)
     }
