+++ src/redisearch_rs/trie_rs/src/iter/mod.rs
@@ -11,8 +11,10 @@
 pub mod filter;
 mod iter_;
 mod lending;
+mod prefixes;
 mod values;
 
 pub use iter_::Iter;
 pub use lending::LendingIter;
+pub use prefixes::PrefixesIter;
 pub use values::Values;
